requirements.txt
scipy
numpy
opencv-python
Pillow
pytesseract
sklearn
six
django
djangorestframework
django-cors-headers
